requirements.txt
# --- CORE INFRASTRUCTURE ---
python-dotenv==1.0.1
pydantic==2.8.2
fastapi==0.111.0
uvicorn==0.30.1
flask==3.0.3

## The Golden Stack (Verified Compatible)
langchain==0.2.11
langchain-community==0.2.10
langchain-core==0.2.23
langchain-experimental==0.0.63
langchain-text-splitters==0.2.2
langgraph==0.1.14
langsmith==0.1.90

# The Critical Fix: Pinning Google to the 0.2 Era
langchain-google-genai==1.0.7

# Drivers
neo4j==5.21.0
google-genai
tiktoken==0.7.0
pypdf==4.3.0
pinecone-client==4.0.0
sentence-transformers==3.0.1

# Multimodal processing
PyMuPDF==1.24.0
Pillow==10.2.0

# Production caching
redis==5.0.1
